Added webhooks controller
Currently, after doing the PayTomorrow application,
if user doesn't click the Close button, we don't update the
order. We rely on the redirect after the close button.
This can result in issues and we do have a way around it.

We'll handle the notify callback that PayTomorrow provides
Ref: https://docs.paytomorrow.com/docs/api-reference/postbacks/payment-pending-postback

Made the webhooks controller a template because each app using this extension
may want to handle the callbacks completely differently based on their
setup